// src/app/modules/admin/customer/report/models/customerlistRptParameter.ts
//import internal from "stream"

export class customerlistRptParameter {
    querytype: string;

    allguest: boolean;
    allday: boolean;

    resstatusidlist: string;
    resstatuslist: string;

    createfrom: string;
    createto: string;
    salesfrom: string;
    salesto: string;
    cinfrom: string;
    cinto: string;
    coutfrom: string;
    coutto: string;
    cancelfrom: string;
    cancelto: string;
    stayfrom: string;
    stayto: string;

    hotelidlist: string;
    hotellist: string;

    marketdetail: boolean;
    marketidlist: string;
    marketlist: string;

    agencydetail: boolean;
    agencyidlist: string;
    agencylist: string;

    boarddetail: boolean;
    boardidlist: string;
    boardlist: string;

    rtdetail: boolean;
    rtidlist: string;
    rtlist: string;

    brtdetail: boolean;
    brtidlist: string;
    brtlist: string;

    natdetail: boolean;
    nationalityidlist: string;
    nationalitylist: string;

    fdetail: boolean;
    taxdetail: boolean;

    rescurridlist: string;
    rescurrlist: string;

    rptcurr: boolean;
    rptcurridlist: string;
    rptcurrlist: string;

    reportId: number;

    rowfieldidlist: string;
    rowfieldlist: string;

    groupfield1idlist: string;
    groupfield1list: string;

    groupfield2idlist: string;
    groupfield2list: string;

    groupfield3idlist: string;
    groupfield3list: string;

    groupfield4idlist: string;
    groupfield4list: string;

    paidpaxidlist: string;
    paidpaxlist: string;

    totalpaxidlist: string;
    totalpaxlist: string;

    postdatefrom: string;
    postdateto: string;

    fposttypeidlist: string;
    fposttypelist: string;

    departmenttypeidlist: string;
    departmenttypelist: string;

    resuuid: string;

    folioid: string;

    acceptfrom: string;
    acceptto: string;
    closefrom: string;
    closeto: string;

    roomid: string;
    unitid: string;
    tasktypeid: string;
    departmentid: string;
    mapping1id: string;
    mapping2id: string;
    mapping3id: string;
    mapping4id: string;
    mapping5id: string;
    mapping6id: string;

    tracecalctype: string;
    export: boolean=false;

    departmentidlist: string;
    departmentlist: string;

    cashieridlist: string;
    cashierlist: string;

    roomTypeslist: string;
    roomlist: string;
    maidlist: string;

    agencygroupdetail: boolean;
    agencygroupidlist: string;
    agencygrouplist: string;

    cheflist:string;
}